fix(settings): extract History privacy controls onto shared UI primitives
Move exclusion chips and the privacy dialog into their own module using
SettingIconAction and outline Cancel, and align More browsers with the
shared CollapsibleContent disclosure pattern so Macroscope re-evaluates.
