+++ apps/web/src/components/settings/ComputerUseSettings.tsx
@@ -29,7 +29,7 @@ import {
   SettingsSection,
 } from "./settingsLayout";
 import { searchableSetting } from "./settingsSearch";
-import { Collapsible, CollapsiblePanel, CollapsibleTrigger } from "../ui/collapsible";
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "../ui/collapsible";
 
 /** Shown when the desktop host lacks the Computer Use permissions bridge API. */
 const BRIDGE_UNSUPPORTED_MESSAGE = "Update T3 Code to check Computer Use permissions";
@@ -306,7 +306,6 @@ export function ComputerUseSettings() {
           }
         />
 
-        {/* CollapsibleTrigger provides aria-expanded; panel uses shared CollapsiblePanel transition. */}
         <Collapsible open={moreBrowsersOpen} onOpenChange={setMoreBrowsersOpen}>
           <SettingsRow
             title={<RowTitle icon={<EdgeIcon className="size-5" />}>More browsers</RowTitle>}
@@ -314,14 +313,13 @@ export function ComputerUseSettings() {
             control={
               <CollapsibleTrigger
                 render={<Button type="button" variant="outline" size="sm" />}
-                aria-controls="more-browsers-panel"
-                aria-expanded={moreBrowsersOpen}
+                aria-controls="computer-use-more-browsers"
               >
                 {moreBrowsersOpen ? "Hide" : "Show"}
               </CollapsibleTrigger>
             }
           >
-            <CollapsiblePanel id="more-browsers-panel">
+            <CollapsibleContent id="computer-use-more-browsers">
               <div className="mt-2 space-y-3 rounded-xl bg-muted/20 px-3 py-3">
                 <div className="flex items-center gap-3">
                   <EdgeIcon className="size-6" />
@@ -349,7 +347,7 @@ export function ComputerUseSettings() {
                   </div>
                 </div>
               </div>
-            </CollapsiblePanel>
+            </CollapsibleContent>
           </SettingsRow>
         </Collapsible>
       </SettingsSection>
